🎺 Add GitHub Actions workflow to build LaTeX document and generate README
This commit adds a GitHub Actions workflow in the `.github/workflows/latex.yml` file. The workflow is triggered on push events and runs on the latest version of Ubuntu. It sets up the Git repository, compiles the LaTeX document using the `xu-cheng/latex-action` action, gets the current date, lists the directory contents, uploads the PDF to a new branch called `book`, and uploads the PDF file as an artifact.

Also, add `.gitignore` file

This commit adds a `.gitignore` file to the repository. The file ignores the `.github/.DS_Store` and `.DS_Store` files.